+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/util/dfu/fixer/v100/NectarDecomponentizeFix.java
@@ -1,4 +1,4 @@
-package house.greenhouse.bovinesandbuttercups.util.dfu.fixer;
+package house.greenhouse.bovinesandbuttercups.util.dfu.fixer.v100;
 
 import com.google.common.collect.ImmutableMap;
 import com.mojang.datafixers.DSL;
@@ -7,7 +7,6 @@ import com.mojang.datafixers.OpticFinder;
 import com.mojang.datafixers.TypeRewriteRule;
 import com.mojang.datafixers.Typed;
 import com.mojang.datafixers.schemas.Schema;
-import com.mojang.datafixers.types.Type;
 import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Dynamic;
 import net.minecraft.util.datafix.fixes.References;