+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/util/dfu/fixer/v101/RenameToAlstroemeriaRanchChunkFix.java
@@ -1,4 +1,4 @@
-package house.greenhouse.bovinesandbuttercups.util.dfu.fixer;
+package house.greenhouse.bovinesandbuttercups.util.dfu.fixer.v101;
 
 import com.mojang.datafixers.DataFix;
 import com.mojang.datafixers.TypeRewriteRule;